rag/requirements.txt
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
google-genai
python-dotenv